requirements.txt
# --- Core UI / system ---
PySide6>=6.6.0
psutil>=5.9.0
pygetwindow>=0.0.9
keyboard>=0.13.5
httpx>=0.27.0

# --- Voice input ---
sounddevice>=0.4.6
faster-whisper>=1.0.0

# --- LLM providers ---
# Cloud provider (free-tier llama-3.3-70b @ ~6000 tok/s).
# Set GROQ_API_KEY env var to enable. Falls back to Ollama if absent.
groq>=0.11.0

# --- TTS providers ---
# Neural TTS (default). Falls back to pyttsx3 if model files missing.
# Download voices: https://github.com/rhasspy/piper/releases
piper-tts>=1.2.0
pyttsx3>=2.90

# --- Calendar (opt-in via config.proactive_enabled) ---
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.115.0

# --- Vector memory (Phase 2A) ---
# Semantic kNN retrieval. Falls back to lexical if either is missing.
sqlite-vec>=0.1.9
fastembed>=0.3.0

# --- Tests ---
pytest>=8.0.0
pytest-cov>=5.0.0

# --- Optional: silero-vad endpointing ---
# Voice listener auto-detects torch and uses silero-vad if available.
# Install separately if you want better endpointing:
#   pip install torch torchaudio
